requirements.txt
pip==20.0.2
Django==2.2.12
npm==0.1.1
django-bootstrap4==1.1.1
djongo==1.3.1
django-crispy-forms==1.9.0
django-background-tasks==1.2.5
tweepy==3.8.0
textblob==0.15.3
MeaningCloud-python==1.2.2